requirements.txt
Django==2.2.10
django-crispy-forms==1.9.2
Pillow==7.2.0
django-debug-toolbar==2.2